ssss

+++ semo22/src/main/java/practice/semo/Member/MemberRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MemberRepository extends JpaRepository<DBMember,Integer> {
+public interface MemberRepository extends JpaRepository<DBMember,Long> {
 
   Optional<DBMember> findByUsername(String username);
